tratamento de exeçao no delete

+++ src/main/java/com/vitor/spring/services/UserService.java
@@ -4,15 +4,18 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.vitor.spring.entities.User;
 import com.vitor.spring.repositories.UserRepository;
+import com.vitor.spring.services.exceptions.DatabaseException;
 import com.vitor.spring.services.exceptions.ResourceNotFoundException;
 
 @Service
 public class UserService {
-    
+
     @Autowired
     private UserRepository userRepository;
 
@@ -20,7 +23,7 @@ public class UserService {
         return userRepository.findAll();
     }
 
-     public User findUserById(Long id) {
+    public User findUserById(Long id) {
         Optional<User> obj = userRepository.findById(id);
         return obj.orElseThrow(() -> new ResourceNotFoundException(id));
     }
@@ -30,12 +33,19 @@ public class UserService {
     }
 
     public void delete(Long id) {
-        userRepository.deleteById(id);
+        if (!userRepository.existsById(id)) {
+            throw new ResourceNotFoundException(id);
+        }
+        try {
+            userRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User obj) {
         User entity = userRepository.getReferenceById(id);
-        updateData(entity,obj);
+        updateData(entity, obj);
         return userRepository.save(entity);
     }
